fix: fix big multiplier being rounded by using new BigNumberInput
The "multiplier" and "evenSegment" values were being set in the UI as numbers.
This was causing an issue were a big multiplier like 11111111111111111 was
being rounded to 11111111111111110, and therefore being treated like an even
number. It was also making multiple parts of the code more complicated.

Now all parts using multiplier and evenSegment expect those variables to have
BigNumber instances.

Other changes:
- set Typescript target to es2017
- add math utility fns isEven and isNumber
- upgraded some dependencies
- noted issue with mathjs v12